Moved load_bakery
